+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 - implemented minimal quiz loop in 'main.py'
 - project runs sucessfully with one random math question
 
-## [v0.11]
+## [v0.1.1]
+- added -QuizManager- class to seperate quiz logic
+- added -QuizInterface- class to handle rendering, currently supports Streamlit